+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.6.3](https://github.com/kubb-project/kubb/compare/@kubb/cli-v1.6.2...@kubb/cli-v1.6.3) (2023-09-03)
+
+
+### 🐞 Bug Fixes
+
+* add `@swc/core` ([7c44878](https://github.com/kubb-project/kubb/commit/7c44878e83d65eeb57dbbea15b2066cc61cc2471))
+
+## [1.6.2](https://github.com/kubb-project/kubb/compare/@kubb/cli-v1.6.1...@kubb/cli-v1.6.2) (2023-09-02)
+
+
+### 🐞 Bug Fixes
+
+* upgrade oas(es support) and overall packages ([c5b1f4e](https://github.com/kubb-project/kubb/commit/c5b1f4e0f6e4fc880df94f8a02d9a0b9b81053ff))
+
 ## [1.6.1](https://github.com/kubb-project/kubb/compare/@kubb/cli-v1.6.0...@kubb/cli-v1.6.1) (2023-08-29)