+++ packages/swagger-zodios/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.6.3](https://github.com/kubb-project/kubb/compare/@kubb/swagger-zodios-v1.6.2...@kubb/swagger-zodios-v1.6.3) (2023-09-03)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/swagger-zodios:** Synchronize undefined versions
+
+## [1.6.2](https://github.com/kubb-project/kubb/compare/@kubb/swagger-zodios-v1.6.1...@kubb/swagger-zodios-v1.6.2) (2023-09-02)
+
+
+### 🐞 Bug Fixes
+
+* support for PATCH ([c326d7c](https://github.com/kubb-project/kubb/commit/c326d7cafaa2ddf258579f889f061bbdf51a96fd))
+
 ## [1.6.1](https://github.com/kubb-project/kubb/compare/@kubb/swagger-zodios-v1.6.0...@kubb/swagger-zodios-v1.6.1) (2023-08-29)