+++ packages/ts-codegen/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.6.3](https://github.com/kubb-project/kubb/compare/@kubb/ts-codegen-v1.6.2...@kubb/ts-codegen-v1.6.3) (2023-09-03)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/ts-codegen:** Synchronize undefined versions
+
+## [1.6.2](https://github.com/kubb-project/kubb/compare/@kubb/ts-codegen-v1.6.1...@kubb/ts-codegen-v1.6.2) (2023-09-02)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/ts-codegen:** Synchronize undefined versions
+
 ## [1.6.1](https://github.com/kubb-project/kubb/compare/@kubb/ts-codegen-v1.6.0...@kubb/ts-codegen-v1.6.1) (2023-08-29)